+++ plone/protect/monkey.py
@@ -1,7 +1,5 @@
-from zope.component.hooks import getSite
 from urlparse import urlparse, urljoin
 from plone.protect.auto import safeWrite
-from Products.PluggableAuthService import utils as pluaggable_utils
 
 
 def RedirectTo__call__(self, controller_state):
@@ -43,17 +41,13 @@ def wl_lockmapping(self, killinvalids=0, create=0):
 
 def pluggableauth__getCSRFToken(request):
     """
-    if we have a site object, let plone.protect do it's job
+    let plone.protect do it's job
     """
-    if getSite():
-        return
-    return pluaggable_utils._old_getCSRFToken(request)
+    return ''
 
 
 def pluggableauth__checkCSRFToken(request, token='csrf_token', raises=True):
     """
-    if we have a site object, let plone.protect do it's job
+    let plone.protect do it's job
     """
-    if getSite():
-        return
-    return pluaggable_utils._old_checkCSRFToken(request)
+    pass